+++ mozrepl/util.py
@@ -3,14 +3,18 @@
 
 from __future__ import unicode_literals, absolute_import, division, print_function
 import cookielib
-#from ufp.terminal.debug import print_ as debug
+from ufp.terminal.debug import print_ as debug
+
+from .exception import Exception as MozException
 
 def convertToJs(arg):
 	"""
 	입력받은 값을 javascript에서 사용 가능한 값으로 변환합니다.
 	
-	:param arg: 변환할 값
-	:type arg: int, float, None, unicode, bytes, str, :py:class:`~mozrepl.type.Object`, :py:class:`~mozrepl.type.Function`, :py:class:`~mozrepl.type.Array`
+	.. todo:: dict 형태 변환 기능을 추가.
+	
+	:param arg: 변환할 값. 만약 list와 tuple타입을 인자로 준다면, 포함된 값은 :py:func:`~mozrepl.util.convertToJs` 함수가 변환 할 수 있는 값이어야 합니다.
+	:type arg: int, float, None, unicode, bytes, str, tuple, list, :py:class:`~mozrepl.type.Object`, :py:class:`~mozrepl.type.Function`, :py:class:`~mozrepl.type.Array`
 	:return: javascript에서 사용 할 수 있는 값.
 	:rtype: unicode
 	"""
@@ -26,23 +30,32 @@ def convertToJs(arg):
 	if isinstance(arg, str):
 		return repr(arg)
 	
-	from ..type import Object
+	if isinstance(arg, (tuple, list)):
+		buffer = map(convertToJs, arg)
+		buffer = ', '.join(buffer)
+		return '[{0}]'.format(buffer)
+	
+	from .type import Object
 	if isinstance(arg, Object):
 		return arg.reference
 	
-	raise TypeError('{}는 처리 할 수 없는 타입입니다.'.format(type(arg)))
+	raise TypeError('"{type}" 타입은 변환 할 수 없는 타입입니다.'.format(type=type(arg)))
 
 def getCookiesFromHost(repl, host):
 	"""
+	host와 일치하는 쿠키를 cookielib.Cookie형식으로 가져옵니다.
+	
+	.. todo:: 버그 있음 수정 할 것.
+	
 	:param repl: mozrepl.Mozrepl 객체
 	:type repl: :py:class:`~mozrepl.Mozrepl`
 	:param host: 호스트
 	:type host: unicode
 	:yield: 각 cookielib.Cookie.
 	"""
-	buffer = repl.execute('Services').cookies.getCookiesFromHost(host)
-	for cookie in buffer:
-		cookie = cookie.QueryInterface(repl.execute('Ci').nsICookie)
+	iter = repl.execute('Services').cookies.getCookiesFromHost(host)
+	while iter.hasMoreElements():
+		cookie = iter.getNext().QueryInterface(repl.execute('Ci').nsICookie)
 		
 		domain = cookie.host
 		initial_dot = domain.startswith(".")